# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dstdelivery",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/get/cart/products": {
            "get": {
                "tags": [
                    "Cart Section"
                ],
                "summary": "Get Cart Products",
                "description": "",
                "operationId": "5109a3eebfcb7725835c2243d3c3910f",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Please write user ID",
                        "required": true,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
        "/api/delete/cart/products": {
            "delete": {
                "tags": [
                    "Cart Section"
                ],
                "summary": "Delete Cart Products",
                "description": "",
                "operationId": "e9a536f744f79a65e07d3584d3a875db",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "Please write product ID",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Please write user ID",
                        "required": true,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
        "/api/add/quantity/forOne/Product": {
            "post": {
                "tags": [
                    "Cart Section"
                ],
                "summary": "Add Quantity For One Product",
                "description": "",
                "operationId": "58a228bce11ba84a2c89a9aa6767c3e7",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "Please write product ID",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Please write user ID",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "totalQty",
                        "in": "query",
                        "description": "Please Write quantity How much do you want to add",
                        "required": true,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
...
        "/api/create/product": {
            "post": {
                "tags": [
                    "Product Section"
                ],
                "summary": "Request that added a new product",
                "description": "",
                "operationId": "18a3477fb98a13fe6f2eade0b8ee6afd",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Provide product title",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Provide product description",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "photoFileName",
                        "in": "query",
                        "description": "Provide product photo file Name",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Provide product status",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "description": "Provide product price",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "totalQty",
                        "in": "query",
                        "description": "Provide product totalQty",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "sub_category_id",
                        "in": "query",
                        "description": "provide category ID or let empty or provide 1 for create default one",
                        "required": true,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
        "/api/update/product": {
            "put": {
                "tags": [
                    "Product Section"
                ],
                "summary": "Request which updating something regarding product",
                "description": "",
                "operationId": "90415969b766028462f70ba471909461",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "Please write product id",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "For example please write new product title",
                        "required": true,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
        "/api/delete/product": {
            "delete": {
                "tags": [
                    "Product Section"
                ],
                "summary": "Request which deletes product",
                "description": "",
                "operationId": "aa859f5491f68a060ce90d736783f763",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "Please write product id",
                        "required": true,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
        "/api/filter/product": {
            "get": {
                "tags": [
                    "Product Section"
                ],
                "summary": "Request that search via name",
                "description": "",
                "operationId": "ed46f17ade888e9b84add373ba7ab169",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Please write product title",
                        "required": true,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
        "/api/get/product/list": {
            "get": {
                "tags": [
                    "Product Section"
                ],
                "summary": "Request which returns categories with products, store, photos and options",
                "description": "",
                "operationId": "de2bbf8ae792edb0f2133652cb265b71",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}